Add createCopies option to disable creating hashed files

+++ tasks/cachebust.js
@@ -10,6 +10,7 @@ var _ = grunt.util._;
 var DEFAULT_OPTIONS = {
     algorithm: 'md5',
     baseDir: './',
+    createCopies: true,
     deleteOriginals: false,
     encoding: 'utf8',
     jsonOutput: false,
@@ -59,9 +60,11 @@ module.exports = function() {
             }));
             var newFilename = addFileHash(file, hash, opts.separator);
 
-            grunt.file.copy(absPath, path.resolve(opts.baseDir, newFilename));
+            if(opts.createCopies) {
+                grunt.file.copy(absPath, path.resolve(opts.baseDir, newFilename));
+            }
 
-            if(opts.deleteOriginals) {
+            if(opts.deleteOriginals && opts.createCopies) {
                 grunt.file.delete(absPath);
             }
